# swagger.yaml
openapi: 3.0.0
info:
  title: TicketFlow API
  version: 1.0.0
  description: API documentation for the TicketFlow event ticketing platform.
servers:
  - url: http://localhost:3001/api # Adjust based on your backend URL
    description: Development server
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Resource not found"
        errors: # Optional array for validation errors
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                example: "body.email"
              message:
                type: string
                example: "Invalid email address"
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: cuid
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        imageUrl:
          type: string
          format: url
          nullable: true
        status:
          type: string
          enum: [DRAFT, PUBLISHED, CANCELLED]
        organizerId:
          type: string
          format: cuid
        organizer: # Included from relation
          type: object
          properties:
            id:
              type: string
              format: cuid
            name:
              type: string
              nullable: true
            email:
              type: string
              format: email
        ticketCategories: # Included from relation
            type: array
            items:
              $ref: '#/components/schemas/TicketCategory'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - date
        - location
        - status
        - organizerId
        - createdAt
        - updatedAt
    TicketCategory:
      type: object
      properties:
        id:
          type: string
          format: cuid
        name:
          type: string
        price:
          type: number
          format: float
        totalQty:
          type: integer
        bookedQty:
          type: integer
        eventId:
          type: string
          format: cuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - totalQty
        - bookedQty
        - eventId
        - createdAt
        - updatedAt
    Booking:
      type: object
      properties:
        id:
          type: string
          format: cuid
        userId:
          type: string
          format: cuid
          nullable: true
        eventId:
          type: string
          format: cuid
        quantity:
          type: integer
        totalPrice:
          type: number
          format: float
        status:
          type: string
          enum: [PENDING, PROCESSING, CONFIRMED, CANCELLED, FAILED]
        deliveryName:
          type: string
          nullable: true
        deliveryEmail:
          type: string
          format: email
          nullable: true
        deliveryPhone:
          type: string
          nullable: true
        utr:
          type: string
          nullable: true
        paymentVerifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        event: # Included from relation
          type: object
          properties:
            id:
              type: string
            title:
              type: string
        user: # Included from relation
          type: object
          properties:
             id:
               type: string
             email:
               type: string
             name:
               type: string
               nullable: true

      required:
        - id
        - eventId
        - quantity
        - totalPrice
        - status
        - createdAt
        - updatedAt

    # ========== Request Bodies ==========
...
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
          example: "Summer Music Festival"
        description:
          type: string
          example: "An amazing weekend of live music."
        date:
          type: string
          format: date-time
          example: "2024-08-15T18:00:00.000Z"
        location:
          type: string
          example: "City Park Amphitheater"
        imageUrl:
          type: string
          format: url
          example: "https://example.com/image.jpg"
        status:
          type: string
          enum: [DRAFT, PUBLISHED, CANCELLED]
          example: DRAFT
        ticketCategories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "General Admission"
              price:
                type: number
                format: float
                example: 50.00
              totalQty:
                type: integer
                example: 1000
            required:
             - name
             - price
             - totalQty
          minItems: 1
      required:
        - title
        - description
        - date
        - location
        - ticketCategories
...
    CreateBookingRequest:
        type: object
        properties:
            eventId:
                type: string
                format: cuid
                example: "clxjq1k2a0000u0pqc9z0h3v7"
            quantity:
                type: integer
                minimum: 1
                example: 2
            deliveryName:
                type: string
                example: "Booker Name"
            deliveryEmail:
                type: string
                format: email
                example: "booker@example.com"
            deliveryPhone:
                type: string
                example: "1234567890"
        required:
            - eventId
            - quantity
            # Make delivery fields optional if collected later
            #- deliveryName
            #- deliveryEmail
            #- deliveryPhone
...
    EventResponse:
        $ref: '#/components/schemas/Event'
    EventsListResponse:
        type: array
        items:
            $ref: '#/components/schemas/Event'
    BookingResponse:
        $ref: '#/components/schemas/Booking'
    BookingsListResponse:
        type: array
        items:
            $ref: '#/components/schemas/Booking'
  securitySchemes:
      bearerAuth:
          type: http
          scheme: bearer
          bearerFormat: JWT
paths:
...
  /events:
    get:
      summary: Retrieve a list of events
      tags: [Events]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, CANCELLED]
          description: Filter events by status
        # Add other query params like date range, location, pagination (page, limit)
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventsListResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Create a new event
      tags: [Events]
      security:
        - bearerAuth: [] # Requires authentication (Admin/Organizer role check in handler)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
           description: Unauthorized (Missing token)
           content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
            description: Forbidden (Invalid token or insufficient role)
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /bookings:
    post:
        summary: Create a new booking
        tags: [Bookings]
        security:
            - bearerAuth: [] # Optional depending on whether anonymous bookings are allowed
        requestBody:
            required: true
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/CreateBookingRequest'
        responses:
            '201':
                description: Booking created successfully
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BookingResponse'
            '400':
                description: Validation Error or Booking Error (e.g., insufficient tickets, event not found)
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ErrorResponse'
            '401':
                description: Unauthorized (If authentication is required)
                content:
                    application/json:
                      schema:
                        $ref: '#/components/schemas/ErrorResponse'
            '500':
                description: Internal Server Error
                content:
                    application/json:
                      schema:
                        $ref: '#/components/schemas/ErrorResponse'

    get:
        summary: Retrieve a list of bookings (User's own or all for Admin)
        tags: [Bookings]
        security:
            - bearerAuth: [] # Requires authentication
        parameters:
          - in: query
            name: status
            schema:
                type: string
                enum: [PENDING, PROCESSING, CONFIRMED, CANCELLED, FAILED]
            description: Filter bookings by status
          - in: query
            name: page
            schema:
                type: integer
                default: 1
            description: Page number for pagination
          - in: query
            name: limit
            schema:
                type: integer
                default: 10
            description: Number of bookings per page
        responses:
            '200':
                description: A list of bookings
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BookingsListResponse'
            '401':
                description: Unauthorized
                content:
                    application/json:
                      schema:
                        $ref: '#/components/schemas/ErrorResponse'
            '500':
                description: Internal Server Error
                content:
                    application/json:
                      schema:
                        $ref: '#/components/schemas/ErrorResponse'
...